Refactor LiquibaseSchemaGeneratorTest into an abstract class
