+++ klass-generators/klass-generator-liquibase-schema-test/src/main/java/cool/klass/generator/liquibase/schema/test/AbstractLiquibaseSchemaGeneratorTest.java
@@ -1,20 +1,4 @@
-/*
- * Copyright 2024 Craig Motlin
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cool.klass.generator.liquibase.schema;
+package cool.klass.generator.liquibase.schema.test;
 
 import cool.klass.model.meta.domain.api.source.DomainModelWithSourceCode;
 import cool.klass.model.meta.loader.compiler.DomainModelCompilerLoader;
@@ -27,9 +11,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 @ExtendWith(LogMarkerTestExtension.class)
-public class LiquibaseSchemaGeneratorTest
+public abstract class AbstractLiquibaseSchemaGeneratorTest
 {
-    public static final String FULLY_QUALIFIED_PACKAGE = "cool.klass.xample.coverage";
+    public static final String FULLY_QUALIFIED_PACKAGE = "replace.with.actual.package";
 
     @RegisterExtension
     final FileMatchExtension fileMatchExtension = new FileMatchExtension(this.getClass());
