+++ klass-generators/klass-generator-liquibase-schema-test/src/main/java/cool/klass/generator/liquibase/schema/test/LiquibaseSchemaGeneratorTest.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2024 Craig Motlin
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cool.klass.generator.liquibase.schema.test;
+
+import cool.klass.model.meta.domain.api.source.DomainModelWithSourceCode;
+import cool.klass.model.meta.loader.compiler.DomainModelCompilerLoader;
+import io.liftwizard.junit.extension.log.marker.LogMarkerTestExtension;
+import io.liftwizard.junit.extension.match.file.FileMatchExtension;
+import org.eclipse.collections.api.list.ImmutableList;
+import org.eclipse.collections.impl.factory.Lists;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.RegisterExtension;
+
+@ExtendWith(LogMarkerTestExtension.class)
+public class LiquibaseSchemaGeneratorTest extends AbstractLiquibaseSchemaGeneratorTest
+{
+    public static final String FULLY_QUALIFIED_PACKAGE = "klass.model.meta.domain";
+
+    @RegisterExtension
+    final FileMatchExtension fileMatchExtension = new FileMatchExtension(this.getClass());
+
+    @Test
+    void smokeTest()
+    {
+        ImmutableList<String> klassSourcePackages = Lists.immutable.with(FULLY_QUALIFIED_PACKAGE);
+
+        var domainModelCompilerLoader = new DomainModelCompilerLoader(
+                klassSourcePackages,
+                Thread.currentThread().getContextClassLoader(),
+                DomainModelCompilerLoader::logCompilerError);
+
+        DomainModelWithSourceCode domainModel = domainModelCompilerLoader.load();
+
+        this.fileMatchExtension.assertFileContents(
+                this.getClass().getCanonicalName() + ".xml",
+                SchemaGenerator.getSourceCode(domainModel, FULLY_QUALIFIED_PACKAGE));
+    }
+}
